+++ packages/deploy-config-sub-generator/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @sap-ux/deploy-config-sub-generator
 
+## 0.4.5
+
+### Patch Changes
+
+-   @sap-ux/abap-deploy-config-sub-generator@0.1.61
+
+## 0.4.4
+
+### Patch Changes
+
+-   @sap-ux/abap-deploy-config-sub-generator@0.1.60
+
 ## 0.4.3
 
 ### Patch Changes