+++ packages/preview-middleware/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @sap-ux/preview-middleware
 
+## 0.23.6
+
+### Patch Changes
+
+-   Updated dependencies [f46dd02]
+    -   @sap-ux/adp-tooling@0.15.25
+
 ## 0.23.5
 
 ### Patch Changes